Change long from int, remove data annotation

+++ src/main/java/engineeringthesis/androidrestapi/game/domain/GameplayRepository.java
@@ -2,6 +2,6 @@ package engineeringthesis.androidrestapi.game.domain;
 
 import engineeringthesis.androidrestapi.common.repository.UUIDAwareJpaRepository;
 
-interface GameplayRepository extends UUIDAwareJpaRepository<Gameplay, Integer> {
+interface GameplayRepository extends UUIDAwareJpaRepository<Gameplay, Long> {
 
 }
